Relative times: minor refactoring
The implementation for relative times has been split into pieces that can be
tested on their own and likewise tests have been added for these.

Now time deltas are calculated independently, and the display messages are only
a reflection of this data, hence allowing to implement any type of messaging
which involves relative times.