Tests: serialize using jsonify in create_api_request
This ensures the same behavior as in application code. Note we don't use a plain
`json.dumps()`, but we provide our own encoding class which supports encoding
more data types than the default implementation.

The conditional import statement is necessary to avoid circular imports and
to allow tests to be run without breaking pytest-django.