RawFontTextarea: integrate undo logic
This is mostly to allow storing selection values in the change history; keeping
such state in a higher level wouldn't allow us to effectively alter selections
on undo and redo actions.
On the other hand, the `Undoable` higher-order component is already coupled with
the editor (via `overrideValue`), so let's not pretend it's a reusable piece.

At the same time, this commit simplifies the amount of state stored in the
undo/redo stack by not keeping track of the current value; note this is already
stored in the editor container. The stack state keys have also been renamed to
`done` and `undone`, as they reflect better what each stack contains.