Use a cleaner way to encode native JSON object
Instead of encoding JSON object as a string and wrapping it
with JSON.parse() call, use a raw JSON string (which
can be embedded into HTML page as is, but make sure
we don't break HTML source parsing by having '</script>' text
inside any JSON string.

This makes resulting HTML+JS code more compact and readable.
I believe this doesn't introduce any security issues, but
this is something that needs to be double-checked for sure
(and we need tests for this, I suppose).