Use Windows-compatible path separator
The way our code currently deals with multiple webpack directories
(as it joins and splits them by ':') doesn't work very well with
Windows paths that have semicolon in their path (C:\path\to\dir).
So changing this to ';' which should be good for both Windows
and *nix.

Also, do not hardcode the trailing forward slash; I believe it is
not necessary in the path.