LastActivity is a future replacement for UserEvent
The reason is that one has to remap all the fields for the data structures
that come from the server in order to use UserEvent. Also, 'UserEvent' name
doesn't reflect the meaning of the component as we call this 'last activity'
(or 'last action') everywhere in the UI and in the data.

This component is currently a shim for UserEvent that provides such mapping
and defaults. When UserEvent is no longer in use, we will merge it with
LastActivity and get rid of UserEvent.js