This is a component to render Stats in a collapsed mode
The component renders three top contributors when stats
are in the collapsed mode. While this component is not
strictly related to the browser table, I'm landing it here
to avoid splitting changes and then dealing with merge
conflicts between separate branches (as both BrowserTable
and StatsCollapsed components require adjacent changes
in stats.js that will be landed later).