Cleanup model field names, get rid of the ones we don't need
This all is an attempt to critically review what is currently
being sent to the client, what is actually needed, to simplify
the rendering and decrease the passed data size at the same time.

For example, we don't need to pass checks_url, profile_url and
various flavors of formatted dates as they can be reconstructed
on the client. Also, unit ID is enough to render last action;
we don't need to pass and cache the entire pootle_path, and can
use short unit permalinks for this purpose.

Also make some field names not too verbose, and unify
displayname/display_name