This helper function cleans up the dictionary
Specifically, it removes default values that should better be
dealt with on the client side (False and zero values, empty strings).
Such a cleanup reduces the noise returned by the server and helps
minimize the traffic.

This commit is not strictly related to the browser table UI,
but the underlying table view code already uses this feature,
so I'm landing it here. It should be easy to remove this function
and a couple of clean_dict() wrapper calls later if desired.