Provide cleaned up / combined stats data to the client
This joins items and stats together, removes unnecessary fields
and sends a cleaned up dictionary into the browser page template.

This by no means is a proper production code I'd be proud of.
Ideally the Python code needs to be refactored to generate
correct data in the first place, but this needs more expertise
than I have in Python/Django.