Tests: implement view testing via snapshots
The concept is simple: request a view via a URL and dump the rendered context to
a JSON file. A human needs to evaluate first whether the rendered context is
valid or not.

Afterwards, the verification process compares the expected context with the
context generated by the test run. If they match, the test passes; fails
otherwise.

To make this possible, tests can use the new `snapshot_stack` fixture, which
provides a context manager implementing a simple stack and returns `Snapshot`
objects in its context. By pushing items (strings) to the snapshot stack, a
unique context can be created with the accumulated values, which then maps to a
file on disk. If a string value in the context stack ends with a forward slash,
it will be treated as a directory, and will be created if it doesn't exist.

Tests will use the `assert_matches()` method on `Snapshot` objects, and provide
the context which the reference snapshot should be compared against. The
comparison works at the string level — contexts are serialized to JSON.

In order to generate snapshot files, `py.test --generate-snapshots` needs to be
run. Snapshot-based tests that don't have files on disk will error out.