Implement keyboard help popup
Related issue that I originally filed in Pootle project:
https://github.com/translate/pootle/issues/5102

I decided to keep some currently non-implemented keys
as commented-out code in the component as a reminder to
what the plan is. We can uncomment them as we implement
additional shortcuts.

The popup will rendered differently on Mac vs everywhere else
(because of different shortcuts)