Use ZingIcons font to render icons
We want our UI to be rendered correctly on hdpi screens.
This commit adds a new font and css file that works
as a drop-in replacement to sprite.css

With this change, we drop old sprite icons and related
`make sprites` command entirely. Note that there's still
some old code (e.g. terminology editor) that technically
references the old icons that this commit removes.
We will need to clean up this obsolete functionality
separately.