TreeItem: return children stats as a list
This was originally a dictionary because the keys were used to query the DOM by
code. With the new table, feeding data is enough.

Note when calculating the browsing UI data, we take advantage of the fact that
children items and stats' items are sorted lists so we can safely combine them,
with no need to query a data source by key.