Browser: ensure consistent data in async endpoint
Previously, cached data from Redis would be used as-is in views. This was the
case both in the browsing view and the internal API endpoint to retrieve stats,
which would be consumed by the client code directly.

With the recent changes in the browsing pages, we changed the shape of the data
used by the client, but we didn't adjust the internal asynchronous endpoint that
retrieves updated stats when these are dirty.

This commit not only fixes that inconsistency, but also enables the endpoint to
return browsing-related data, not only stats. This data will be the same as the
data provided in templates' context.

Fixes #62.