Project: remove treestyle field
It is out of the scope of Zing to accommodate to all types of directory layouts
and as such, it will always use what it used to be known before as non-GNU
layout. All projects are meant to follow this structure and it is up to external
localization processes to provide files with the layout adapted to it.