Navbar: improve menu and sign in button positioning on mobile
Previously, when using floating div, the menu would jump to the bottom
(and essentially disappear) on sertain small resolutions. Because of this,
on mobile there would be no way to access the 'sign in' button.

This commit changes the behavior to use an absolutely positioned div
that is always on top of the other navbar elements.