Tests: use fake Redis client for the 'stats' cache
This almost entirely removes the dependency on a Redis server for testing.
Unfortunately `fakeredis` cannot be used for the 'redis' named cache (used for
revisions), as django-redis' `incr` implementation uses Redis' `eval()` command,
which is unsupported in `fakeredis` for the time being.

The situation might change when django-redis moves away from using `eval()` or
`fakeredis` provides support for `eval()`.