Tests: avoid modifying module variable from setup
When using the `request_users` and `site_users` fixtures, it becomes non-obvious
that `TEST_USERS` may contain anything more than the attributes defined in
place. Nevertheless, the current implementation is problematic if we want to
perform the environment setup via a dump, as the fixtures would never have the
`user` key that some tests are expecting.