Tests: restore data dumps when setting up environment
The step to bootstrap the DB with initial data is quite slow, so we can keep a
data dump for further restore via Django's `loaddata`. Besides improving speed,
this also helps us have a consistent baseline to run tests against and
afterwards implement snapshot testing.

Note loading the data alone has some caveats, and hence for instance the
collation of specific columns needs to be fixed before loading the data. Also
note changes to the DB schema will require to rebuild the data dump (remove the
file, re-run the test suite and it will be recreated).