Tests: ensure stats can be made available
We need to make sure tests that require stats data have it available to them.
Refreshing stats after session-level DB setup is not a viable option, because as
tests are executed the stats data gets modified, and since Redis doesn't have
rollback capabilities, the only way to ensure a clean slate for a test requiring
stats data is to calculate these right before running specific test functions.

Since test functions can be parametrized things can get tricky, because tests
sharing the same parameter (e.g. a user) will be run next to each other before
continuing with the next parameter. This means that there still will be risk of
poisoning caches before the test is run with the next parameter.

This commit on the one hand makes the `refresh_stats` fixture available to test
functions. This fixture will ensure to flush the stats cache and repopulate it
with fresh data.

On the other hand, the `test_browse` test function has been adapted to use this
fixture. However, because of the problem mentioned above, the fixture needs to
be requested conditionally: only when we switched users across test runs.

This is quite unconvenient and the parameter to depend on can vary from test to
test. Ideally we shouldn't be concerned about such chances for poisoning (and we
avoid that by making the `TreeItem` implementation a no-op by default in tests),
although let's consider this solution as an interim workaround.