Tests: add empty directories
This commit adds two types of empty directories:

  * A directory which contains no stores at all, hence it is marked as
    `obsolete` according to our current importing logic.
  * A directory which contains a store with no units at all. In practical terms
    this directory is empty for users, although our system doesn't mark it as
    obsolete.

Note we need to manually mark the first type of empty directory as obsolete to
imitate file scanning performed via `update_stores`. Without this, the situation
wouldn't be similar to a real world scenario, the directory would remain alive
and it would screw stats.

For the future it would definitely help to have a better way to initialize these
factories while ensuring the test suite's consistency with a production system,
otherwise we will need to "manually" handle potential differences as in this
case.