Editor: after last unit always show the lead-out message
Previously lead-out 'congratulations!' message would be shown
depending on the number of rows. Now this message is shown
only when the user goes past the last row (closes the editor)
and is accompanied with the same 'Go back to browsing' link
in case there are no more units to go through.