APIView: support permission handlers
Before actual dispatching happens to handler methods, a permission
check will be done based on the `permission_classes` classes.

Permission classes should implement a `has_permission` method, receiving the
current `request` and the `view` which is performing the check. The method will
return a boolean value indicating whether the request is allowed or not.