APIView: support object-specific permission handlers
Whenever individual objects want to be retrieved, permission handlers will have
a chance to determine whether the request has permission to continue further in
the `has_object_permission` method.

This method works the same way as `has_permission`, but it also receives the
object as the third argument. The return value is the same too, a boolean
indicating whether the request is allowed or not.