Due dates: display pending tasks in language browsing contexts
Registered users will have a list of tasks pending for a language
whenever they are in any language-specific browsing page. By default the
top 3 tasks are displayed, but more can be loaded on demand.

The importance of tasks is determined by calculating its critical
factor, which prioritizes the pending work according to the amount of
days left and giving importance to fixing critical errors first.