TreeItem: items have a single parent
This commit renames the `get_parents()` method into `get_parent()` and
returns individual objects instead of lists of objects. Likewise, `None`
is returned for the situation where there are no parents for an item.

Note this very same change was already performed in the past (82740dd),
but the change was reverted to support virtual folders (7a66eae).