Stats: provide a simple way to retrieve stats for a path
This stems from the existing code in `CachedTreeItem`, but keeps it
simple and lightweight, as it requires no DB model instance at all.

With the new `Stats` model, passing a string path is enough to be able
to easily query the Redis stats DB.

As a future thought, we could compose existing Django models with
instances of this class instead of relying in the implicity imposed by
the (`Cached`)`TreeItem` mixin classes.