Due dates: refresh list of pending tasks upon modifications
Whenever due dates are updated from the browsing pages, the list of
tasks visible at the top will be refreshed.

This operation involves the `DueDateContainer` and
`PendingTaskContainer` components, however these currently share no
common ancestor because they are rendered independently. Therefore,
in this commit communication across components happens via the
`PTL.stats` module, however it should be noted that this is a temporal
solution and we should merge the afore-mentioned containers so that
everything is rendered below the same rendering tree.