DueDatePicker: fix day selection in timezones negative wrt UTC
The due date picker was using ISO-formatted date strings as part of its
state, which were then used to create `Date` objects to be used in the
`DatePicker` component.

This turned out to be problematic because `new Date('YYYY-MM-dd')` sets
a date at midnight time in UTC. At the same time, react-day-picker works
with local times internally (e.g. `date.getDate()` instead of
`date.getUTCDate()`), and feeding dates set at UTC midnight time to the
library would result in the day before the selected one being displayed.

So instead of keeping track of a `YYYY-MM-dd` date in the component's
state, we store the date object provided by react-day-picker. At
initialization, the timestamp value is used to generate a `Date` object.
The `YYYY-MM-dd` format is kept as-is for communicating changes to the
server.