Convert failing checks table into a component
The internal API changed a bit, so now it returns not only the check
code and count, but a few more metadata needed to render the table.
Previously this was rendered as part of the HTML on every page load, so
now the browsing context gets slimmer.

This refactoring will help us in the transition to make the statistics
panel a component with a single root container, hence avoiding to have
multiple individual mount points here and there as we have now.
Likewise, this will help to keep and share state across components that
don't have an immediately common ancestor.