add_correction: fix pytz timezone offset issue
It turns out that `get_default_timezone()` can return a timezone used in
ancient times, because this is being handled by pytz and it uses the
Olson timezone DB. There is a detailed explanation of pytz's behavior at
https://bugs.launchpad.net/pytz/+bug/1394441

The bottom line is that using Python datetime's API to deal with
`tzinfo` attributes can result buggy. Therefore, we must use pytz's own
APIs to ensure the right timezone is assigned when localizing times.
Note this is done via Django's own helpers, which handle a few details
for us.