DueDatePicker: fix display of timezones
Whenever a day is selected via react-day-picker, the `day` object
provided by the library is set to `12:00:00` in the client's local time.
In order to dynamically display the proper server time the selected day
represents, the client-side date needs to be:

  * Adjusted to the server's timezone
  * Set to 9:00 in the server's timezone

This requires knowledge of the UTC offset in the client, as well as some
timezone arithmetic which is now done via a couple utility functions.

Implementation note: this hasn't been implemented as a specialized
subclass of `Date` (e.g. `ServerDate`) because subclassing of native
types is limited to pure ES6 environments, and hence transpiling
wouldn't allow us to get there.
Refs. https://babeljs.io/docs/usage/caveats/#classes