Dropdown: convert into controlled component
This commit makes the dropdown component into a controlled one so that
state can be managed from outside. This is useful when willing to
open/close/toggle the dropdown based on some external state.

At the same time, the API has been modified to split it into three
required pieces:

* `<Dropdown />` serves as a wrapper for the contents. It takes care
  of setting up event listeners for closing the dropdown via generic
  methods (escape key press, mouse click on document) and also forwarding
  down the context which is used by the other two components. An
  `onToggle` prop needs to be passed in order to control the open/close
  state upon clicking the dropdown toggle.
* `<DropdownToggle />` sets up the toggling functionality and the
  children elements will be used as the trigger targets.
* `<DropdownMenu />` encloses the contents that will be displayed once
  the dropdown has been opened.

Note that in order to enable this declarative API, the context feature
has been purposefully used. This is similar to a `<select>` + `<option>`
API, where components need to implicitly pass information around.
Refs. https://twitter.com/dan_abramov/status/749715530454622208

Last but not least, `<Dropdown />` and `<DropdownToggle />` also accept
a `tag` prop to specify the wrapper tag to be used. This will be useful
to render dropdown wrappers as inline elements, as well as to use
buttons or other components as triggers.