update_stores: ensure obsolete and unsynced units can be resurrected
When calculating the diff for the update operation, it needs to be
determined which units to use to compare them against active DB units.
If there were unsynced changes in the DB, the algorithm was taking out
DB units updated since the last sync which are obsolete, hence making it
impossible to resurrect them.

At the same time, even if we properly determine which units to compare
to, the attempt to update obsolete unsynced units in
`StoreUpdater.update_unit()` would fail, because it would:
  1. State that there is a conflict
  2. Add a suggestion for the obsolete unit

Which means the unit wouldn't go through the unit-level merge logic in
`Unit.update()`, and therefore the unit wouldn't be resurrected.

This could be fixed differently, either by
  a) resurrecting conflictive obsolete units in
    `StoreUpdater.update_unit()`, or
  b) determining that obsolete units are not conflictive, hence making
    it go through the unit-level merge logic, which would then resurrect
    the unit.

We decided on a) because it is consistent behavior with respect to
updates done when there are unsynced changes in non-obsolete state, and
requires no extra efforts from translators.

Fixes #246.