Invoice tests: use already-existing submission
It turns out that adding a new submission at the time of creating a
Submission via its factory (required to create a ScoreLog) would trigger
creating a new ScoreLog as a post-save operation. This would then lead
to unexpected numbers when testing. So we workaround that by providing
an existing (arbitrary) Submission object to the ScoreLog factory.