SocialVerificationView: ensure the right form is used
When the sign in view was switched to be customized via Allauth's own
form overridding method in 38f7e27, that made `SignInForm` to be used
all the time.

This can be explained by checking the parent Allauth-provided
`LoginView`'s logic to retrieve the form class to use:

> return get_form_class(app_settings.FORMS, 'login', self.form_class)

which will always give preference to the overriden login form specified
in the settings.

This commit modifies `SocialVerificationView` so that we ensure
`SocialVerificationForm` will always be used.