Reports.add_correction: use server timezone explicitly
When no timezone is specified, `timezone.localtime()` attempts to use
the *current* active timezone which is relevant in the context of web
requests, however when using it in a CLI context the active timezone
wouldn't be found and it would fall back to the *default* (server)
timezone, making it work properly in an accidental/implicit way.

In order to avoid potential issues in the future, we'll pass the
timezone explicitly so that we know the server timezone is being used.