Tests: update snapshots
In Django 1.11, form widgets are rendered using templates, and the
output of attributes of HTML tags are not sorted, which means the output
is not deterministic. This is why the `search_form` context attribute
needs to be taken out.

Likewise, widgets now inject extra information as part of the template
rendering context, which is reflected in the snapshot changes.