Django 1.11: widget's `format_output()` is deprecated
We don't need to replace it by a template-based widget, because
rendering is performed in the client.

For the same reasons, the `HiddenWidget` is not needed anymore; it's not
rendered in templates.

We keep the widget because it is able to compress/uncompress plural
values (we may want to do this solely as part of the field in the
future).