Docs: port to Docusaurus
This commit moves the documentation from Sphinx to Docusaurus, hence
allowing us to write docs in Markdown format.

Docusaurus' feature-set is more limited in comparison to Sphinx, however
its simplicity is enough to cover our needs. Apart from that, writing
docs and contributing fixes becomes easier.

The generated docs not only serve as a documentation site (which, by the
way, can easily and comfortably be browsed on GitHub), but also as a
landing page and general-purpose website.

Likewise, contents have been overhauled, removing irrelevant parts and
focusing on simplicity. There's still work to do here and not all of the
contents are entirely up-to-date, but this is a huge change with respect
to what we had.