Cleanup: rename ctx variable for clarity
The idea is the context variable must indicate if a correction already
exists for that particular month (i.e. if some correction has already
been recorded). The previous naming made it unclear whether it referred
to a particular invoice run generating a correction or simply stating
that a particular month contains a correction. For the sake of ctx data
and display, we are looking for the latter, hence the rename.