Invoices: fix reporting of carried over balance
The previous implementation had a few flaws, the main problem being that
invoice amounts were being calculated every time when access to such
data was needed. This happened from multiple call sites, and provided
inconsistent results based on whether a correction was added or not.

Now amounts are only calculated at the time the invoice is generated,
making the whole thing more predictable, and totals are accessible via
`self.amounts`. In other words, in order to access the invoice data
`generate()` must be called first (there are assertions to ensure this
is the case) and this, in turn, will create and adjust corrections as
needed.

The amounts data will now as well indicate the relevant correction
information.