Cleanup: clarify corrections and carry-overs
To make it clearer the naming has been renamed to `is_carried_over` and
`_add_carry_over`.

An invoice can have *multiple corrections*, hence saying it has a
correction makes it not really clear when we want to indicate that an
amount wasn't paid in the month being processed and it'll be carried
over to the next month.