Omit microsecond precision in date/time utils
There might be problems with MySQL versions 5.6 and newer because for
columns without microsecond precision, passing data with microseconds
will end up in rounding, as opposed to truncation (which was the default
in versions 5.5 and older).

Servers can always adjust column types to be microsecond-aware, although
this can be pretty easy to go unnoticed, therefore we will skip
microseconds altogether because we don't need them and they can cause
the afore-mentioned problems.

Refs. https://bugs.mysql.com/bug.php?id=68760