Due dates: exclude project paths in `for_project_path`
Because apart from language-specific paths this was returning project
paths, updating/deleting would acquire locks for such paths, but that
was already being modified (and hence locked) as part of the same
transaction when calling the `save()` and `delete()` methods of parent
classes.

Therefore excluding the project path from this method (which
semantically was incorrect) should fix the locking issue.