Checks: avoid breakage with language-specific checks
The list of available quality checks doesn't include language-specific
checks, and with that, the information provided for the quality check is
incomplete (e.g. missing `is_critical` field).

Given TTK's nature, retrieving language-specific check information would
be way too complicated for the benefit it'd provide (especially given
that there are no language-specific checks which are critical), and
hence we'll simply provide the checks `name` and the `title` unchanged,
and use a `is_critical` value of `False`.