py3: make sure serialization returns bytes
Previously the `__str__` method would perform the conversion step, and
the `serialize` method was relying on that behavior. When we merged
`__unicode__` and `__str__` methods though, that behavior ceased to
exist.

At the same time, we need to serialize to bytes (which was str in PY2),
so that change is needed too.