Tests: change Redis cache DB location
This is to avoid clashes with existing local configurations. There was
an issue where running tests would flush the entire Redis DB (which then
would require an expensive recalculation!). While this shouldn't happen
in practice because we use fakeredis, there seems to be some
misconfiguration on our end. Let's revisit that later — this changeset
should avoid issues in the meantime.